Avoid short/ushort in MSL code generator.
The Metal code generator has historically avoided low-precision types in
the final output in order to dodge a variety of type-coercion issues.
However, the workaround was only coded for half/float. Extended the
workaround to cover int/short and uint/ushort as well.

Reviewed-on: https://skia-review.googlesource.com/c/skia/+/437688
Commit-Queue: John Stiles <johnstiles@google.com>
Auto-Submit: John Stiles <johnstiles@google.com>

+++ tests/sksl/shared/ArrayIndexTypes.metal
@@ -10,8 +10,8 @@ fragment Outputs fragmentMain(Inputs _in [[stage_in]], bool _frontFacing [[front
     Outputs _out;
     (void)_out;
     array<float, 4> array = array<float, 4>{1.0, 2.0, 3.0, 4.0};
-    short x = 0;
-    ushort y = 1u;
+    int x = 0;
+    uint y = 1u;
     int z = 2;
     uint w = 3u;
     _out.sk_FragColor = float4(array[x], array[y], array[z], array[w]);
